requirements.txt
Django==4.2.6
Pillow==10.1.0
asgiref==3.7.2
djangorestframework==3.14.0
pip==23.2.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
python-decouple==3.8